requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-core
faiss-cpu
ollama
python-dotenv
beautifulsoup4
langsmith
